--- src/main/java/com/booster/cinemagic/controlador/CarteleraControlador.java
@@ -1,84 +1,0 @@
-package com.booster.cinemagic.controlador;
-
-import com.booster.cinemagic.excepciones.DatosInvalidosExcepcion;
-import com.booster.cinemagic.excepciones.ElementoNoEncontradoExcepcion;
-import com.booster.cinemagic.excepciones.IdNoValidoExcepcion;
-import com.booster.cinemagic.modelos.entidad.CarteleraEntidad;
-import com.booster.cinemagic.servicio.ICarteleraServicio;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.List;
-
-@RestController
-@RequestMapping("/cinemagic/v0/cartelera")
-public class CarteleraControlador {
-    /*
-
-    @Autowired
-    ICarteleraServicio carteleraServicio;
-
-    @GetMapping()
-    public List<CarteleraEntidad> consultarcarteleras() {
-        return carteleraServicio.obtenerCarteleras();
-    }
-
-    @GetMapping("/{id}")
-    public CarteleraEntidad consultarCarteleraPorId(@PathVariable Integer id) {
-        CarteleraEntidad cartelera;
-        if(id < 1){
-            throw new IdNoValidoExcepcion("El id debe ser numerico y mayor 0", "Error 410", HttpStatus.BAD_REQUEST);
-        }
-
-        cartelera = carteleraServicio.obtenerCarteleraPorId(id);
-        if(cartelera == null){
-            throw new ElementoNoEncontradoExcepcion("No se encontro ninguna cartelera con ese id:" + id,"Error 409",HttpStatus.NOT_FOUND);
-        }
-        return cartelera;
-    }
-
-    @PostMapping()
-    public CarteleraEntidad agregarCartelera(@Valid @RequestBody CarteleraEntidad cartelera, BindingResult resultado) {
-        if(resultado.hasErrors()){
-            throw new DatosInvalidosExcepcion("Error de validacion - Favor de revisar los datos introducidos", "Error 408",HttpStatus.BAD_REQUEST, resultado);
-        }
-        return carteleraServicio.agregarCartelera(cartelera);
-    }
-
-    @PutMapping("/{id}")
-    public CarteleraEntidad modificarCartelera(@PathVariable Integer id, @Valid  @RequestBody CarteleraEntidad carteleraModificada, BindingResult resultado) {
-        CarteleraEntidad cartelera;
-        if(id < 1){
-            throw new IdNoValidoExcepcion("El id debe ser numerico y mayor 0", "Error 410",HttpStatus.BAD_REQUEST);
-        }
-        if(resultado.hasErrors()){
-            throw new DatosInvalidosExcepcion("Error de validacion - Favor de revisar los datos introducidos", "Error 408",HttpStatus.BAD_REQUEST, resultado);
-        }
-        cartelera = carteleraServicio.modificarCartelera(id,carteleraModificada);
-        if(cartelera == null){
-            throw new ElementoNoEncontradoExcepcion("No se encontro ninguna cartelera con ese id:" + id,"Error 409",HttpStatus.NOT_FOUND);
-        }
-
-        return cartelera;
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> eliminarCartelera(@PathVariable Integer id){
-        boolean respuesta;
-        if(id < 1){
-            throw new IdNoValidoExcepcion("El id debe ser numerico y mayor 0", "Error 410",HttpStatus.BAD_REQUEST);
-        }
-        respuesta = carteleraServicio.eliminarCartelera(id);
-        if(!respuesta){
-            throw new ElementoNoEncontradoExcepcion("No se encontro ninguna cartelera con ese id:" + id,"Error 409",HttpStatus.NOT_FOUND);
-        }
-
-        return ResponseEntity.ok("Se eliminó la cartelera con id: "+ id);
-    }
-
-     */
-}
