+++ src/main/java/com/booster/cinemagic/filtros/JwtSolicitudFiltro.java
@@ -1,0 +1,55 @@
+package com.booster.cinemagic.filtros;
+
+import com.booster.cinemagic.servicios.autentificacion.DetallesUsuarioServicio;
+import com.booster.cinemagic.servicios.autentificacion.JwtUtilService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+
+@Component
+public class JwtSolicitudFiltro extends OncePerRequestFilter {
+
+    @Autowired
+    private DetallesUsuarioServicio detallesUsuarioServicio;
+
+    @Autowired
+    private JwtUtilService jwtUtilService;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+        final String authorizationHeader = request.getHeader("Authorization");
+
+        String username = null;
+        String jwt = null;
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            jwt = authorizationHeader.substring(7);
+            username = jwtUtilService.extractUsername(jwt); //extrae username de jwt
+        }
+
+        if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+
+            UserDetails userDetails = this.detallesUsuarioServicio.loadUserByUsername(username);
+
+            if (jwtUtilService.validateToken(jwt, userDetails)) {
+
+                UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+                        userDetails, null, userDetails.getAuthorities());
+                authenticationToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                SecurityContextHolder.getContext().setAuthentication(authenticationToken);
+            }
+        }
+        filterChain.doFilter(request, response);
+    }
+}
