--- src/main/java/com/booster/cinemagic/repositorio/ICarteleraRepositorio.java
@@ -1,10 +1,0 @@
-package com.booster.cinemagic.repositorio;
-
-import com.booster.cinemagic.modelos.entidad.CarteleraEntidad;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ICarteleraRepositorio {
-    //extends JpaRepository<CarteleraEntidad, Integer>
-}
