--- src/main/java/com/booster/cinemagic/repositorio/ISalaRepositorio.java
@@ -1,9 +1,0 @@
-package com.booster.cinemagic.repositorio;
-
-import com.booster.cinemagic.modelos.entidad.SalaEntidad;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ISalaRepositorio extends JpaRepository<SalaEntidad, Integer> {
-}
