+++ src/main/java/com/booster/cinemagic/repositorios/ICarteleraRepositorio.java
@@ -1,0 +1,8 @@
+package com.booster.cinemagic.repositorios;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ICarteleraRepositorio {
+    //extends JpaRepository<CarteleraEntidad, Integer>
+}
