--- src/main/java/com/booster/cinemagic/servicio/ICarteleraServicio.java
@@ -1,20 +1,0 @@
-package com.booster.cinemagic.servicio;
-
-import com.booster.cinemagic.modelos.entidad.CarteleraEntidad;
-
-import java.util.List;
-
-public interface ICarteleraServicio {
-    /*
-    List<CarteleraEntidad> obtenerCarteleras();
-
-    CarteleraEntidad obtenerCarteleraPorId(Integer id);
-
-    CarteleraEntidad agregarCartelera(CarteleraEntidad cartelera);
-
-    CarteleraEntidad modificarCartelera(Integer id, CarteleraEntidad cartelera);
-
-    boolean eliminarCartelera(Integer id);
-
-     */
-}
