--- src/main/java/com/booster/cinemagic/servicio/IEstadoServicio.java
@@ -1,19 +1,0 @@
-package com.booster.cinemagic.servicio;
-
-import com.booster.cinemagic.modelos.entidad.EstadoEntidad;
-
-import java.util.List;
-
-
-public interface IEstadoServicio {
-
-    List<EstadoEntidad> obtenerEstados();
-
-    EstadoEntidad obtenerEstadoPorId(Integer id);
-
-    EstadoEntidad agregarEstado(EstadoEntidad estado);
-
-    EstadoEntidad modificarEstado(Integer id, EstadoEntidad estado);
-
-    boolean eliminarEstado(Integer id);
-}
