+++ src/main/java/com/booster/cinemagic/servicios/ICarteleraServicio.java
@@ -1,0 +1,16 @@
+package com.booster.cinemagic.servicios;
+
+public interface ICarteleraServicio {
+    /*
+    List<CarteleraEntidad> obtenerCarteleras();
+
+    CarteleraEntidad obtenerCarteleraPorId(Integer id);
+
+    CarteleraEntidad agregarCartelera(CarteleraEntidad cartelera);
+
+    CarteleraEntidad modificarCartelera(Integer id, CarteleraEntidad cartelera);
+
+    boolean eliminarCartelera(Integer id);
+
+     */
+}
