+++ src/main/java/com/booster/cinemagic/servicios/ISalaServicio.java
@@ -1,0 +1,18 @@
+package com.booster.cinemagic.servicios;
+
+import com.booster.cinemagic.entidadades.SalaEntidad;
+
+import java.util.List;
+
+public interface ISalaServicio{
+
+    List<SalaEntidad> obtenerSalas();
+
+    SalaEntidad obtenerSalaPorId(Integer id);
+
+    SalaEntidad agregarSala(SalaEntidad sala);
+
+    SalaEntidad modificarSala(Integer id, SalaEntidad sala);
+
+    boolean eliminarSala(Integer id);
+}
