+++ src/main/java/com/booster/cinemagic/util/TokenInfoModelo.java
@@ -1,0 +1,14 @@
+package com.booster.cinemagic.util;
+
+public class TokenInfoModelo {
+
+    private final String jwtToken;
+
+    public TokenInfoModelo(String jwtToken) {
+        this.jwtToken = jwtToken;
+    }
+
+    public String getJwtToken() {
+        return this.jwtToken;
+    }
+}
